app/requirements.txt
fastapi
sqlalchemy
passlib[bcrypt]
python-dotenv
pydantic
uvicorn
pymysql
python-multipart